+++ src/AuthTelegrambotUserRepository.php
@@ -8,6 +8,7 @@ abstract class AuthTelegrambotUserRepository extends BaseRepository
 {
     /**
      * @param string $_telegram_id
+     *
      * @return AuthTelegrambotUserEntity|null
      */
     abstract public function getByTelegramId(string $_telegram_id): ?AuthTelegrambotUserEntity;